Added instructions: ADD, ADC, SUB, SBC

+++ include/cpu.hpp
@@ -75,8 +75,8 @@ public:
     // Stack operations
     void stack_push(u8 value);
     void stack_push16(u16 value);
-    u16 stack_pop();
-    u8 stack_pop16();
+    u8 stack_pop();
+    u16 stack_pop16();
     u16 stack_peek();
 
 private:
